[docs] create EditScreensCardInner for separete copmonent in EditFamily and EditHouseworkList
