+++ src/screens/EditFamily.jsx
@@ -1,45 +1,16 @@
 import React, { useState } from 'react'
 import { useHistory } from 'react-router-dom'
-import { useMediaQuery } from 'react-responsive'
 import { makeStyles, createMuiTheme } from '@material-ui/core/styles'
-import {
-  Card,
-  ListSubheader,
-  Avatar,
-  CardContent,
-  Typography,
-  List,
-  IconButton,
-  Container,
-  Button,
-  Divider,
-  CardActionArea,
-} from '@material-ui/core'
-import DeleteIcon from '@material-ui/icons/Delete'
+import { ListSubheader, List, Container, Button } from '@material-ui/core'
 import AddIcon from '@material-ui/icons/Add'
 import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos'
 import { ThemeProvider } from '@material-ui/styles'
 
 import EditMember from './EditMember'
 import AddMember from './AddMember'
+import EditScreensCardInner from '../components/EditScreensCardInner'
 
 const useStyles = makeStyles(() => ({
-  card: {
-    display: 'flex',
-    maxWidth: 600,
-    width: '100%',
-    height: '10%',
-    margin: 10,
-  },
-  content_area: {
-    display: 'flex',
-    alignItems: 'center',
-    width: '100%',
-    padding: 0,
-    '&:last-child': {
-      paddingBottom: 0,
-    },
-  },
   container: {
     flexGrow: 1,
     maxWidth: 600,
@@ -49,36 +20,16 @@ const useStyles = makeStyles(() => ({
     justifyContent: 'space-around',
     padding: '0 5%',
   },
-  img: {
-    marginRight: 16,
-  },
-  img_sp: {
-    marginRight: 8,
-    width: 20,
-    height: 20,
-  },
-  txtbox: {
-    width: '100%',
-    display: 'flex',
-  },
-  card_txt: {
-    lineHeight: 3,
-    verticalAlign: 'middle',
-    paddingLeft: 16,
-  },
-  divider: {
-    margin: '0 4px',
-  },
-  btn_back: {
-    marginTop: 8,
-    width: 30,
-    fontSize: 16,
-  },
   btn_icon: {
     margin: 0,
   },
   btn: {
     margin: 24,
+  },
+  btn_back: {
+    marginTop: 8,
+    width: 30,
+    fontSize: 16,
   },
 }))
 //theme color change
@@ -86,9 +37,6 @@ const theme = createMuiTheme({
   palette: {
     primary: {
       main: '#6587ae',
-    },
-    secondary: {
-      main: '#765759',
     },
   },
 })
@@ -100,10 +48,6 @@ const EditFamily = ({ membersInfo, updateFirestoreMember, addFiestoreMember, del
   console.log({ membersInfo })
   const classes = useStyles()
   const history = useHistory()
-
-  //media query
-  const isTablet = useMediaQuery({ query: '(min-device-width: 768px)' })
-  const isSmartPhone = useMediaQuery({ query: '(max-device-width: 767px)' })
 
   const handleBackHome = () => {
     console.log({ history })
@@ -130,6 +74,11 @@ const EditFamily = ({ membersInfo, updateFirestoreMember, addFiestoreMember, del
     setIsEdit(false)
     setIsAdd(false)
   }
+
+  //card内で表示する要素
+  const displayElmNameSp = ['name', 'point']
+  const displayElmName = ['name', 'point']
+
   //メンバー編集画面、メンバー追加画面はflagを使って対応させる
   if (isAdd && !isEdit) {
     return (
@@ -174,89 +123,31 @@ const EditFamily = ({ membersInfo, updateFirestoreMember, addFiestoreMember, del
               家族編集
             </ListSubheader>
           }
-          className={classes.root}
         >
           {membersInfo.map((memberInfo, i) => {
             return (
-              <Card className={classes.card} key={i}>
-                <CardActionArea
-                  // memberInfo={memberInfo}
-                  // handleSubmitMember={(member) => handleSubmitMember(member)}
-                  // handleSubmitMember={handleSubmitMember}
-                  onClick={() => handleEditMember(i)}
-                >
-                  <CardContent className={classes.txtbox}>
-                    {isSmartPhone && (
-                      <CardContent className={classes.content_area}>
-                        <Avatar className={classes.img_sp} variant="circular" />
-                        <Typography
-                          align="left"
-                          display="inline"
-                          // gutterBottom
-                          variant="subtitle2"
-                        >
-                          {memberInfo.name}
-                        </Typography>
-                      </CardContent>
-                    )}
-
-                    {isTablet && (
-                      <CardContent className={classes.content_area}>
-                        <Avatar className={classes.img} variant="circular" />
-                        <Typography
-                          align="left"
-                          display="inline"
-                          // gutterBottom
-                          variant="h5"
-                          component="p"
-                        >
-                          {memberInfo.name}
-                        </Typography>
-                        <Typography
-                          className={classes.card_txt}
-                          display="inline"
-                          variant="body1"
-                          color="textSecondary"
-                          component="p"
-                        >
-                          {memberInfo.level} Lv
-                        </Typography>
-                        <Typography
-                          className={classes.card_txt}
-                          display="inline"
-                          variant="body1"
-                          color="textSecondary"
-                          component="p"
-                        >
-                          {memberInfo.point}Point
-                        </Typography>
-                      </CardContent>
-                    )}
-                  </CardContent>
-                </CardActionArea>
-                <ThemeProvider theme={theme}>
-                  <Divider className={classes.divider} orientation="vertical" flexItem />
-                  <IconButton
-                    color="secondary"
-                    onClick={() => deleteFirestoreMember(memberInfo.id)}
-                    aria-label="delete"
-                  >
-                    <DeleteIcon />
-                  </IconButton>
-                </ThemeProvider>
-              </Card>
+              <EditScreensCardInner
+                data={memberInfo}
+                displayElmNameSp={displayElmNameSp}
+                displayElmName={displayElmName}
+                i={i}
+                handleEdit={handleEditMember}
+                handleDelete={deleteFirestoreMember}
+              />
             )
           })}
         </List>
-        <Button
-          variant="contained"
-          onClick={() => setIsAdd(true)}
-          color="primary"
-          className={classes.btn}
-          startIcon={<AddIcon />}
-        >
-          家族を追加する
-        </Button>
+        <ThemeProvider theme={theme}>
+          <Button
+            variant="contained"
+            onClick={() => setIsAdd(true)}
+            color="primary"
+            className={classes.btn}
+            startIcon={<AddIcon />}
+          >
+            家族を追加する
+          </Button>
+        </ThemeProvider>
       </Container>
     )
   }
