+++ src/uikit/BackBtn.jsx
@@ -1,25 +1,16 @@
 import React from 'react'
-import { makeStyles } from '@material-ui/core/styles'
 import { Button } from '@material-ui/core'
 import ArrowBackIosIcon from '@material-ui/icons/ArrowBackIos'
-const useStyles = makeStyles(() => ({
-  btn_back: {
-    marginTop: 8,
-    width: 30,
-    fontSize: 16,
-  },
-}))
 
-const BackBtn = ({ handleBack }) => {
-  const classes = useStyles()
+const BackBtn = ({ className, handleBack }) => {
   return (
     <Button
+      className={className}
       variant="text"
       color="inherit"
       size="small"
-      className={classes.btn_back}
       onClick={() => handleBack()}
-      startIcon={<ArrowBackIosIcon className={classes.btn_icon} />}
+      startIcon={<ArrowBackIosIcon />}
     >
       back
     </Button>
